--- a/L/LLVM/LLVM_full@19/build_tarballs.jl
+++ b/L/LLVM/LLVM_full@19/build_tarballs.jl
@@ -1,7 +1,7 @@
-version = v"19.1.1"
+version = v"19.1.7"
 
 include("../common.jl")
 
 build_tarballs(ARGS, configure_build(ARGS, version; experimental_platforms=true)...;
-               preferred_gcc_version=v"10", preferred_llvm_version=v"18", julia_compat="1.6")
+               preferred_gcc_version=v"10", preferred_llvm_version=v"16", julia_compat="1.6")
 # Build trigger: 7