--- a/L/LLVM/LLVM_full@19/build_tarballs.jl
+++ b/L/LLVM/LLVM_full@19/build_tarballs.jl
@@ -4,4 +4,4 @@ include("../common.jl")
 
 build_tarballs(ARGS, configure_build(ARGS, version; experimental_platforms=true)...;
                preferred_gcc_version=v"10", preferred_llvm_version=v"16", julia_compat="1.6")
-# Build trigger: 7
+# Build trigger: 8