--- a/L/LLVM/common.jl
+++ b/L/LLVM/common.jl
@@ -20,7 +20,7 @@ const llvm_tags = Dict(
     v"16.0.6" => "499f87882a4ba1837ec12a280478cf4cb0d2753d", # julia-16.0.6-2
     v"17.0.6" => "0007e48608221f440dce2ea0d3e4f561fc10d3c6", # julia-17.0.6-5
     v"18.1.7" => "ed30d043a240d06bb6e010a41086e75713156f4f", # julia-18.1.7-2
-    v"19.1.7" => "6e8ff5767e9466d73be85f1eaa760e2c94cdbd86", # julia-19.1.7-0
+    v"19.1.7" => "3cb75bacf7f61380a4ef45fef519c2fb5940dd4a", # julia-19.1.7-0
 )
 
 const buildscript = raw"""