Also switch clang back to v16 for compilation as that seems to be faster.