@eschnett 2025-02-09
> Also switch clang back to v16 for compilation as that seems to be faster.

How did you measure this? This wouldn't have involved ccache, would it?