@giordano 2025-02-09
> How did you measure this? This wouldn't have involved ccache, would it?

That's still relevant for the first run with cold cache, and for windows we always use clang and disable ccache: https://github.com/JuliaPackaging/Yggdrasil/blob/332689dad13f1b6a0bf101f9445d954fd65e7111/L/LLVM/common.jl#L51-L56 https://github.com/JuliaPackaging/Yggdrasil/blob/332689dad13f1b6a0bf101f9445d954fd65e7111/L/LLVM/common.jl#L286-L286