@Zentrik 2025-02-09
> > Also switch clang back to v16 for compilation as that seems to be faster.
> 
> How did you measure this? This wouldn't have involved ccache, would it?

When binary builder added clang 17, my builds all started timing out (i.e taking hours longer) on the CI at https://github.com/Zentrik/llvm_julia_tester. I don't think I even had ccache setup then. 
I suspect I tested the speed of clang 18 similarly at some point.