@Zentrik 2025-02-09
> @Zentrik are we good to build on [JuliaLang/llvm-project@3cb75ba](https://github.com/JuliaLang/llvm-project/commit/3cb75bacf7f61380a4ef45fef519c2fb5940dd4a) now (remember to bump the triggers to force rebuild for all platforms)? We should have reasonably warm ccache and multi-threaded auditor, hopefully this should save some time during the builds.

Hopefully, I have a build going on at https://github.com/Zentrik/llvm_julia_tester/actions/runs/13227764722 to confirm. But if building here won't require much work from you feel free to kick it off.