@giordano 2025-02-09
[Previous build, LLVM_full i686-linux-gnu-cxx11](https://buildkite.com/julialang/yggdrasil/builds/17427#0194e68e-abef-463f-b740-23cf7e60e1c3/6-8516):
```
[ Info: Timings: setup: 1m 44.77s, build: 40m 16.86s, audit: 70m 47.38s, packaging: 53.28s
```
[New build, LLVM_full i686-linux-gnu-cxx11](https://buildkite.com/julialang/yggdrasil/builds/17480#0194ece4-3aef-4631-9860-2d2236f02226/6-7738):
```
[ Info: Timings: setup: 2m 57.87s, build: 8m 25.5s, audit: 33m 39.8s, packaging: 30.32s
```
Parallel audit id finally bearing fruits.