@Zentrik 2025-02-10
> Good to go once last job finishes?

Yep 🥳