@giordano 2025-02-10
Well, I don't know if we'll ever get out of the race condition on
```
[08:58:28] /workspace/srcdir/llvm-project/llvm/include/llvm/CodeGenTypes/MachineValueType.h:43:10: fatal error: 'llvm/CodeGen/GenVT.inc' file not found
[08:58:28] #include "llvm/CodeGen/GenVT.inc"
[08:58:28]          ^~~~~~~~~~~~~~~~~~~~~~~~
[08:58:28] 1 error generated.
```
We hit it three times in a row in the remaining job...